Move ReviewFlagFormSet and related code to editors module
It's only used there anyway.
